made several stuff public to make this working without needing additional setter methods
